requirements.txt
streamlit>=1.25
pdfplumber
python-docx
spacy
rapidfuzz
requests
reportlab
